(#1113) support compound filters in inventory discovery
